Added eventmanager and removed code from gameengine
